requirements.txt
pandas
numpy
scikit-learn
flask
flask-cors
mlflow
pyYAML
tqdm
ensure
joblib
types-PyYAML
python-box
matplotlib
notebook
ipykernel
python-dotenv